+++ src/main/resources/edu/cornell/kfs/concur/cu-spring-concur.xml
@@ -174,7 +174,7 @@
                                     <bean parent="DelimitedFlatFilePropertySpecification" p:lineSegmentIndex="182" p:propertyName="cashAdvanceIssueDate"
                                     	p:formatterClass="org.kuali.kfs.sys.businessobject.format.BatchDateFormatter"
                                     	p:dateFormat="yyyy-MM-dd"/>
-                                    <bean parent="DelimitedFlatFilePropertySpecification" p:lineSegmentIndex="186" p:propertyName="cashAdvanceCaKey"/>
+                                    <bean parent="DelimitedFlatFilePropertySpecification" p:lineSegmentIndex="186" p:propertyName="cashAdvanceKey"/>
                                     <bean parent="DelimitedFlatFilePropertySpecification" p:lineSegmentIndex="190" p:propertyName="chartOfAccountsCode"/>
                                     <bean parent="DelimitedFlatFilePropertySpecification" p:lineSegmentIndex="191" p:propertyName="accountNumber"/>
                                     <bean parent="DelimitedFlatFilePropertySpecification" p:lineSegmentIndex="192" p:propertyName="subAccountNumber"/>
